Разделил файл на 3 отдельных (по классу).
